Chaining Selectors Exercise
Added properties to elements on index page. Made the element with both the `avatar` and `proportioned` classes 300 pixels wide, then give it a height so that it retains its original square proportions. Made the element with both the `avatar` and `distorted` classes 200 pixels wide, then make its height twice as big as its width.